# packages/scanner/tests/fixtures/shopapi.openapi.json
{"openapi":"3.1.0","info":{"title":"ShopAPI","description":"Demo e-commerce API for local security testing only.","version":"1.0.0"},"paths":{"/login":{"post":{"tags":["auth"],"summary":"Login","description":"Authenticate with email/password and return a bearer token.\n\n# VULN: no rate limiting — this endpoint accepts unlimited attempts\n# and is therefore brute-forceable.","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["users"],"summary":"Get User","description":"Return a user record.\n\nAuth is required, but any authenticated caller may request any id.","operationId":"get_user_users__user_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["users"],"summary":"Update User","description":"Update a user profile, including privileged fields from the body.","operationId":"update_user_users__user_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders":{"get":{"tags":["orders"],"summary":"List My Orders","description":"List orders for the authenticated caller only.\n\nThis endpoint is *correctly scoped* so the scanner must distinguish\nit from the BOLA object-level route below.","operationId":"list_my_orders_orders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderOut"},"type":"array","title":"Response List My Orders Orders Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders/{order_id}":{"get":{"tags":["orders"],"summary":"Get Order","description":"Return a single order by id.","operationId":"get_order_orders__order_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products":{"get":{"tags":["products"],"summary":"List Products","description":"Return the public catalog. No auth, no sensitive fields.","operationId":"list_products_products_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProductOut"},"type":"array","title":"Response List Products Products Get"}}}}}}},"/products/search":{"get":{"tags":["products"],"summary":"Search Products","description":"Search the catalog by product name.\n\n# VULN: SQL injection — `q` is concatenated into a raw SELECT. A payload\n# such as q=' OR '1'='1 bypasses the name filter and returns every row.\n# SELECT-only: this handler never UPDATE / DELETE / DROP.","operationId":"search_products_products_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductOut"},"title":"Response Search Products Products Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/stats":{"get":{"tags":["admin"],"summary":"Admin Stats","description":"Return internal store metrics.\n\n# VULN: broken authentication — no auth dependency, reachable by anyone.","operationId":"admin_stats_admin_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminStatsOut"}}}}}}},"/healthz":{"get":{"tags":["ops"],"summary":"Healthz","description":"Liveness probe used by run scripts and later orchestration.","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthOut"}}}}}}}},"components":{"schemas":{"AdminStatsOut":{"properties":{"total_users":{"type":"integer","title":"Total Users"},"total_orders":{"type":"integer","title":"Total Orders"},"total_revenue":{"type":"number","title":"Total Revenue"},"pending_chargebacks":{"type":"integer","title":"Pending Chargebacks"}},"type":"object","required":["total_users","total_orders","total_revenue","pending_chargebacks"],"title":"AdminStatsOut","description":"Internal store metrics."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthOut":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"}},"type":"object","required":["status","service"],"title":"HealthOut","description":"Liveness payload."},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest","description":"Email/password login body."},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["access_token","user_id"],"title":"LoginResponse","description":"Bearer token issued after a successful login."},"OrderOut":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"product_id":{"type":"integer","title":"Product Id"},"quantity":{"type":"integer","title":"Quantity"},"total":{"type":"number","title":"Total"},"status":{"type":"string","title":"Status"},"shipping_address":{"type":"string","title":"Shipping Address"}},"type":"object","required":["id","user_id","product_id","quantity","total","status","shipping_address"],"title":"OrderOut","description":"Order detail, including owner id and shipping address."},"ProductOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"price":{"type":"number","title":"Price"}},"type":"object","required":["id","name","description","price"],"title":"ProductOut","description":"Public catalog product — no sensitive fields."},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"password_hash":{"type":"string","title":"Password Hash"},"is_admin":{"type":"boolean","title":"Is Admin"},"phone":{"type":"string","title":"Phone"},"internal_notes":{"type":"string","title":"Internal Notes"}},"type":"object","required":["id","email","password_hash","is_admin","phone","internal_notes"],"title":"UserOut","description":"User profile as returned by the API.\n\nIncludes fields a correct API would never return to a peer customer."},"UserUpdate":{"properties":{"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"is_admin":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Admin","description":"Privileged flag. A correct API must ignore this from clients."}},"type":"object","title":"UserUpdate","description":"Profile update body.\n\n`is_admin` is accepted on purpose so the mass-assignment detector\ncan see the privileged field in the OpenAPI spec."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}